+++ protocols/identify/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Add `hide_listen_addrs` option to prevent leaking (local) listen addresses.
   See [PR 5507](https://github.com/libp2p/rust-libp2p/pull/5507).
+- Add `with_cache_config` to fully configure the identify cache using `PeerAddressesConfig`.
+  Deprecating `with_cache_size` in favor of `with_cache_config`.
+  See [PR 5574](https://github.com/libp2p/rust-libp2p/pull/5574).
 
 ## 0.45.0