+++ protocols/identify/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.45.2
+
+- Make stream timeout and maximum number of concurrent streams configurable on identify.
+  See [PR 5654](https://github.com/libp2p/rust-libp2p/pull/5654).
+
 ## 0.45.1
 
 - Add `hide_listen_addrs` option to prevent leaking (local) listen addresses.